backend/requirements.txt

fastapi
uvicorn[standard]
pydantic-settings
pymupdf
python-multipart
numpy
psycopg2-binary
google-genai
tenacity
